auth.requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
psycopg2
